enhances code readability and adds TSDoc to "removeEmptyFolders"

+++ src/index/formatFileStructure/removeEmptyFolders.ts
@@ -1,18 +1,27 @@
 import fs from "fs";
 import path from "path";
 
-export const removeEmptyFolders = (p: string) => {
-  const files = fs.readdirSync(p);
-  if (!files) {
-    fs.rmdirSync(p);
+/**
+ * @summary
+ * Recursively removes all empty folders.
+ *
+ * @param directory the full path of a directory
+ *
+ * @example
+ * removeEmptyFolders('/home/user/documents/project/src')
+ */
+export function removeEmptyFolders(directory: string): void {
+  const files = fs.readdirSync(directory);
+  if (!files) return fs.rmdirSync(directory);
+
+  for (const base of files) {
+    const fullPath = path.resolve(directory, base);
+    const isDirectory = fs.lstatSync(fullPath).isDirectory();
+
+    if (!isDirectory) continue;
+
+    removeEmptyFolders(fullPath);
+    const isEmpty = !fs.readdirSync(fullPath).length;
+    if (isEmpty) fs.rmdirSync(fullPath);
   }
-  for (const file of files) {
-    const fullPath = path.join(p, file);
-    if (fs.lstatSync(fullPath).isDirectory()) {
-      removeEmptyFolders(fullPath);
-      if (!fs.readdirSync(fullPath).length) {
-        fs.rmdirSync(fullPath);
-      }
-    }
-  }
-};
+}
